requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
idna==3.3
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
rsa==4.8
six==1.16.0
urllib3==1.26.10
